Progressed HyAtom module to working state

+++ include/QMechModel.h
@@ -1,7 +1,7 @@
 #ifndef QMECH_MODEL_H
 #define QMECH_MODEL_H
 
-#include "../dependencies/glm/glm.hpp"
+#include <glm/glm.hpp>
 
 namespace PhysConstants {
    constexpr double BOHR_R = 1.0;
@@ -10,9 +10,9 @@ namespace PhysConstants {
 
 namespace WaveFunction {
 
-   double computeProbabilityDensity (const glm::vec3& pos, int n, int l, int m);
+   double computeProbabilityDensity (const glm::dvec3& pos, int n, int l, int m);
 
-   double computePhase (const glm::vec3& pos, int n, int l, int m);
+   double computePhase (const glm::dvec3& pos, int n, int l, int m);
 
 };
